Add BufferReader tests
